requirements.txt
Django==1.6.4
South==0.8.4
django-admin-bootstrapped==1.6.4
django-cors-headers==0.12
django-filter==0.7
djangorestframework==2.3.13
wsgiref==0.1.2